API documentation

Construct and work with a CiteLibrary

CitableLibrary.library — Function

Construct a CiteLibrary.

library(collections; libname, liburn, license, cexversion)

Optional parameters

liburn Unique identifier for this library. Must be a subtype of Urn.

libname Name of library (value returned by label)

license License. Default is cc-nc-by.

cexversion Version of the CEX standard used in serializing library. Default is the current version.

CitableLibrary.collectiontypes — Function

List types of collections in library.

collectiontypes(lib)

CitableLibrary.collections — Function

List all collections in the library.

collections(lib)

List all collections in the library of type T.

collections(lib, T)

CitableLibrary.license — Function

License for collected material in library as a whole.

license(lib)

CitableLibrary.cexversion — Function

CEX version to use in serialization.

cexversion(lib)

Functions operating on citable collections

CitableLibrary.citablecollection — Function

True if x has the value CitableLibraryCollection for the CitableLibraryTrait.

citablecollection(x)

Internals

CitableLibrary.UuidUrn — Type

A UUID value in URN notation.

CitableLibrary.uuidUrn — Function

Generate a new UuidUrn

uuidUrn()