API documentation

Construct and work with a CiteLibrary

CitableLibrary.citeLibrary — Function

Construct a CiteLibrary.

citeLibrary(collections; libname, liburn, license, cexversion)

CitableLibrary.collectiontypes — Function

List types of collections in library.

collectiontypes(lib)

CitableLibrary.collections — Function

List all collections in the library.

collections(lib)

List all collections in the library of type T.

collections(lib, T)

Serialize a library

CitableBase.cex — Function

Delegate cex to specific functions based on type's citable trait value.

cex(x; delimiter)

It is an error to invoke the cex function on material that is not citable.

cex(, x; delimiter)

Citable text content should implement cex.

cex(, txt; delimiter)

Citable content should implement cex.

cex(, obj; delimiter)

Delegate cex to specific functions based on type's CitableLibraryTrait value.

cex(urn, x)

It is an error to invoke the cex function with collections that are not citable.

cex(_, urn, x)

Serialize contents of lib to a CEX string.

cex(lib; delimiter)

Functions operating on citable collections

CitableBase.urncontains — Function

Urn subtypes should implement urncontains(urn1::U, urn2::U)::Bool

CitableLibrary.urnequals — Function

Delegate urnequals to specific functions based on type's CitableLibraryTrait value.

It is an error to invoke the urnequals function with collections that are not citable.

Apply urnequals to all collections in lib of type T.

urnequals(urn, lib, T)

Returns a possibly empty list of objects.

Base.iterate — Function

iterate(iter [, state]) -> Union{Nothing, Tuple{Any, Any}}

Advance the iterator to obtain the next element. If no elements remain, nothing should be returned. Otherwise, a 2-tuple of the next element and the new iteration state should be returned.

CitableLibrary.citable — Function

True if x has the value CitableLibraryCollection for the CitableLibraryTrait.

True if x has the value CitableLibraryCollection for the CitableLibraryTrait.

citable(_, x)

Internals

CitableLibrary.UuidUrn — Type

A UUID value in URN notation.

CitableLibrary.uuidUrn — Function

Generate a new UuidUrn

uuidUrn()